Completed desired functionality of Morass Web.
Nodes have more responsiblity than standard networks due to the way pulsing works. The intention is to allow for total parallelisation with the current structure down to individual nodes being parallelised.

Genetic algorithm has been implemented with an example application (digits of pi). Pi may not be a very good application of the Morass Web, but some digits are learned. The next learning task will be for sine waves, which I believe will work much better as the problem domain is small and cyclical..

Signed-off-by: Matthew Muller <matthew.muller@uon.edu.au>